+++ Sources/TunnelKitOpenVPNManager/OpenVPN+ProviderConfiguration.swift
@@ -59,6 +59,9 @@ extension OpenVPN {
         /// The client configuration.
         public let configuration: OpenVPN.Configuration
         
+        /// Enables kill switch.
+        public var killSwitch: Bool?
+
         /// The optional username.
         public var username: String?
         
@@ -95,7 +98,7 @@ extension OpenVPN {
 
 extension OpenVPN.ProviderConfiguration: NetworkExtensionConfiguration {
 
-        public func asTunnelProtocol(
+    public func asTunnelProtocol(
         withBundleIdentifier tunnelBundleIdentifier: String,
         extra: NetworkExtensionExtra?
     ) throws -> NETunnelProviderProtocol {
@@ -112,6 +115,9 @@ extension OpenVPN.ProviderConfiguration: NetworkExtensionConfiguration {
         }
         protocolConfiguration.disconnectOnSleep = extra?.disconnectsOnSleep ?? false
         protocolConfiguration.providerConfiguration = try asDictionary()
+        if #available(iOS 14, *), let killSwitch = killSwitch {
+            protocolConfiguration.includeAllNetworks = killSwitch
+        }
         return protocolConfiguration
     }
 }